requirements.txt
streamlit
streamlit_qrcode_scanner
gspread
pandas
google-auth
gspread
psycopg2-binary